requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
streamlit>=1.32.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pillow>=10.2.0
pandas>=2.2.0
scikit-learn>=1.4.0
streamlit-drawable-canvas>=0.9.3 